requirements.txt
pyirsdk
requests
pyinstaller